+++ ShopperInsights/src/main/java/com/braintreepayments/api/shopperinsights/ShopperInsightsClient.kt
@@ -5,6 +5,7 @@ import com.braintreepayments.api.core.AnalyticsEventParams
 import com.braintreepayments.api.core.AnalyticsParamRepository
 import com.braintreepayments.api.core.BraintreeClient
 import com.braintreepayments.api.core.BraintreeException
+import com.braintreepayments.api.core.DeviceInspector
 import com.braintreepayments.api.core.ExperimentalBetaApi
 import com.braintreepayments.api.core.MerchantRepository
 import com.braintreepayments.api.core.TokenizationKey
@@ -13,6 +14,7 @@ import com.braintreepayments.api.shopperinsights.ShopperInsightsAnalytics.GET_RE
 import com.braintreepayments.api.shopperinsights.ShopperInsightsAnalytics.GET_RECOMMENDED_PAYMENTS_SUCCEEDED
 import com.braintreepayments.api.shopperinsights.ShopperInsightsAnalytics.PAYPAL_PRESENTED
 import com.braintreepayments.api.shopperinsights.ShopperInsightsAnalytics.PAYPAL_SELECTED
+import com.braintreepayments.api.shopperinsights.ShopperInsightsAnalytics.VENMO_PRESENTED
 import com.braintreepayments.api.shopperinsights.ShopperInsightsAnalytics.VENMO_SELECTED
 
 /**
@@ -31,14 +33,18 @@ class ShopperInsightsClient internal constructor(
         EligiblePaymentsApi(braintreeClient, analyticsParamRepository)
     ),
     private val merchantRepository: MerchantRepository = MerchantRepository.instance,
+    private val deviceInspector: DeviceInspector = DeviceInspector(),
+    private val shopperSessionId: String? = null
 ) {
 
     /**
      * @param context: an Android context
      * @param authorization: a Tokenization Key or Client Token used to authenticate
+     * @param shopperSessionId: the shopper session ID returned from your server SDK request
      */
-    constructor(context: Context, authorization: String) : this(
-        BraintreeClient(context, authorization)
+    constructor(context: Context, authorization: String, shopperSessionId: String? = null) : this(
+        BraintreeClient(context, authorization),
+        shopperSessionId = shopperSessionId
     )
 
     /**
@@ -59,7 +65,10 @@ class ShopperInsightsClient internal constructor(
         analyticsParamRepository.resetSessionId()
         braintreeClient.sendAnalyticsEvent(
             GET_RECOMMENDED_PAYMENTS_STARTED,
-            AnalyticsEventParams(experiment = experiment)
+            AnalyticsEventParams(
+                experiment = experiment,
+                shopperSessionId = shopperSessionId
+            )
         )
 
         if (request.email == null && request.phone == null) {
@@ -135,7 +144,7 @@ class ShopperInsightsClient internal constructor(
         callback: ShopperInsightsCallback,
         error: Exception
     ) {
-        braintreeClient.sendAnalyticsEvent(GET_RECOMMENDED_PAYMENTS_FAILED)
+        braintreeClient.sendAnalyticsEvent(GET_RECOMMENDED_PAYMENTS_FAILED, analyticsParams)
         callback.onResult(ShopperInsightsResult.Failure(error))
     }
 
@@ -145,7 +154,7 @@ class ShopperInsightsClient internal constructor(
         isPayPalRecommended: Boolean,
         isVenmoRecommended: Boolean,
     ) {
-        braintreeClient.sendAnalyticsEvent(GET_RECOMMENDED_PAYMENTS_SUCCEEDED)
+        braintreeClient.sendAnalyticsEvent(GET_RECOMMENDED_PAYMENTS_SUCCEEDED, analyticsParams)
         callback.onResult(
             ShopperInsightsResult.Success(
                 ShopperInsightsInfo(
@@ -168,6 +177,7 @@ class ShopperInsightsClient internal constructor(
      * @param pageType optional Represents the page or view the button is rendered on.
      */
     fun sendPresentedEvent(
+        shopperSessionId: String? = null,
         presentmentDetails: PresentmentDetails? = null,
         paymentMethodsDisplayed: List<String> = emptyList(),
         buttonType: ButtonType? = null,
@@ -179,6 +189,7 @@ class ShopperInsightsClient internal constructor(
             AnalyticsEventParams(
                 experiment = presentmentDetails.toString(),
                 paymentMethodsDisplayed = paymentMethodsDisplayed,
+                shopperSessionId = shopperSessionId,
                 buttonType = buttonType.toString(),
                 buttonOrder = buttonOrder.toString(),
                 pageType = pageType.toString()
@@ -191,15 +202,55 @@ class ShopperInsightsClient internal constructor(
      * This method sends analytics to help improve the Shopper Insights feature experience.
      */
     fun sendPayPalSelectedEvent() {
-        braintreeClient.sendAnalyticsEvent(PAYPAL_SELECTED)
+        braintreeClient.sendAnalyticsEvent(PAYPAL_SELECTED, analyticsParams)
     }
+
+//    /**
+//     * Call this method when the Venmo button has been successfully displayed to the buyer.
+//     * This method sends analytics to help improve the Shopper Insights feature experience.
+//     *
+//     * @param experiment optional JSON string representing an experiment you want to run
+//     * @param paymentMethodsDisplayed optional The list of available payment methods,
+//     * rendered in the same order in which they are displayed
+//     */
+//    fun sendVenmoPresentedEvent(
+//        experiment: String? = null,
+//        paymentMethodsDisplayed: List<String> = emptyList()
+//    ) {
+//        braintreeClient.sendAnalyticsEvent(
+//            VENMO_PRESENTED,
+//            AnalyticsEventParams(
+//                experiment = experiment,
+//                paymentMethodsDisplayed = paymentMethodsDisplayed,
+//                shopperSessionId = shopperSessionId
+//            )
+//        )
+//    }
 
     /**
      * Call this method when the Venmo button has been selected/tapped by the buyer.
      * This method sends analytics to help improve the Shopper Insights feature experience.
      */
     fun sendVenmoSelectedEvent() {
-        braintreeClient.sendAnalyticsEvent(VENMO_SELECTED)
+        braintreeClient.sendAnalyticsEvent(VENMO_SELECTED, analyticsParams)
+    }
+
+    /**
+     * Indicates whether the PayPal App is installed.
+     */
+    fun isPayPalAppInstalled(context: Context): Boolean {
+        return deviceInspector.isPayPalInstalled(context)
+    }
+
+    /**
+     * Indicates whether the Venmo App is installed.
+     */
+    fun isVenmoAppInstalled(context: Context): Boolean {
+        return deviceInspector.isVenmoInstalled(context)
+    }
+
+    private val analyticsParams: AnalyticsEventParams get() {
+        return AnalyticsEventParams(shopperSessionId = shopperSessionId)
     }
 
     companion object {